% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.R
\docType{data}
\name{env_name}
\alias{env_name}
\title{The environment name (internal)}
\format{An object of class \code{character} of length 1.}
\usage{
env_name
}
\description{
The environment name (internal)
}
\keyword{datasets}
