% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.R
\name{make_persist_env}
\alias{make_persist_env}
\title{Create the persistent environment to store the persisted variable (internal)}
\usage{
make_persist_env()
}
\value{
The environment used by the package
}
\description{
Create the persistent environment to store the persisted variable (internal)
}
