% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.R
\name{persist}
\alias{persist}
\title{Make a variable persistent}
\usage{
persist(a_var)
}
\arguments{
\item{a_var}{The variable to make persisence}
}
\description{
Make a variable persistent and remove it from interactive scope.
The persistent variables will not be cleaned by the \code{rm(list = ls())}.
}
\details{
The persisted variables are session-specific. They will be gone after you restart R

Note that if you define a new variable with the same name as the persisted variable,
the new variable will mask the persisted one until it is removed.
}
\examples{
x <- "hello"
persist(x)
# x is now not in the .GlobalEnv but still accessible
}
