% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.R
\name{make_persist_env}
\alias{make_persist_env}
\title{Get or create the persistent environment to store the persisted variable}
\usage{
make_persist_env(env_name = "__cleanslate")
}
\value{
the environment used by the package
}
\description{
Get or create the persistent environment to store the persisted variable
}
